Group implementation
Note: the Go database session model uses the hex for both
source and group_id. Did not check the Java version. This
might be in case SQLite does not want bytes in the table,
as little sense as that would make because usually SQLite
does not care about anything.

That's the implementation duplicated here, although the test
case data emulates having proper bytes as the group's id.