Message handler
Although `process_message` needs to mutate `new_message`
after resolving a `session_id` for it, we can transfer
ownership of messages with mutability enabled.

The `NewMessage.session_id` is changed to `Option` to
avoid using an integer for a canary value. Leveraging the
type system makes more sense.

The receipt handler is to be done in another commit.