Require UUID and mime_guess libraries
UUID should be [fastuuid](https://github.com/pradovic/fastuuid-rs)
to match Go, but it doesn't work without a [faster-hex](https://github.com/nervosnetwork/faster-hex)
with the `sse4.1` feature, which doesn't exist for arm.

There is a [PR open](https://github.com/nervosnetwork/faster-hex/pull/13)
but in the finest tradition of free software, we don't know when
it'll be merged...

... and `fastuuid` adamantly installs an older `faster-hex`
regardless of what we do.

The `mime_guess` library is used to deduce file extensions by
mime types, though some of them need to be hard-coded before any
guesses are taken.