Update message's int types to match Signal service
This issue was swept under the rug because Diesel
doesn't like unsigned integers and `i32` and `i64`
worked just fine.

The likely reason this ended up this way in the
first place is that the Go code never complained,
so we could sort of not think about it until now.