Load emojis from $HOME or fall back to system emojis
The emoji system is initialized when WF is started. The parse()
function checks if the heart emoji (u2764) can be loaded and falls
back to the system font if it fails. The check is run only once.

New source path: ~/.local/share/harbour-whisperfish/emojis
Instead of: /usr/share/harbour-whisperfish/icons/emojis