Replace SqliteConnection lock with a standard mutex
The Reentrant lock of parking_lot does not allow for &mut references,
nor would it make sense to artificially add them via a RefCell.

This complicates the access to the database, since we cannot reenter
anymore.  Most of the reentering code is probably fixed by this patch,
except for some calls that use a transaction and then re-enter using
&Storage methods.

These cases can probably be alleviated by moving the remaining logic
outside of the transaction.  It usually purely consists of a final
select logic.